+++ src/Orc.DataAccess.Example/App.xaml.cs
@@ -60,11 +60,9 @@ public partial class App
 
         StyleHelper.CreateStyleForwardersForDefaultStyles();
 
-#if NETCORE
         var factory = System.Data.SqlClient.SqlClientFactory.Instance;
         DbProviderFactories.RegisterFactory("System.Data.SqlClient", factory);
         DbProviderFactories.RegisterFactory("Microsoft.Data.SqlClient", SqlClientFactory.Instance);
-#endif
 
         base.OnStartup(e);
     }
