+++ src/Orc.DataAccess.Xaml/Controls/DbProviderPicker/ViewModels/DbConnectionProviderListViewModel.cs
@@ -1,11 +1,10 @@
 ﻿namespace Orc.DataAccess.Controls;
 
 using System.Collections.Generic;
-using System.Data.Common;
 using System.Linq;
 using System.Threading.Tasks;
+using Catel;
 using Catel.MVVM;
-using CsvHelper;
 using Database;
 
 public class DbConnectionProviderListViewModel : ViewModelBase
@@ -22,7 +21,7 @@ public class DbConnectionProviderListViewModel : ViewModelBase
         DbProviders = new List<DbProviderInfo>();
     }
 
-    public override string Title => "Select provider";
+    public override string Title => LanguageHelper.GetRequiredString(nameof(Properties.Resources.Controls_DbConnectionProviderList_Title));
     public DbProviderInfo? DbProvider { get; set; }
     public IList<DbProviderInfo> DbProviders { get; private set; }
     public Command Refresh { get; }
@@ -49,13 +48,5 @@ public class DbConnectionProviderListViewModel : ViewModelBase
     {
         DbProviders = Database.DbProvider.GetRegisteredProviders().Select(x => x.Value.Info).ToList();
         DbProvider = DbProviders.FirstOrDefault(x => x.Equals(_selectedProvider));
-    }
-
-    private void On()
-    {
-        Orc.DataAccess.Database.
-            DbProvider.UnregisterProvider(new DbProviderInfo(string.Empty, "Microsoft.Data.SqlClient", string.Empty, string.Empty));
-
-        DbProviderFactories.UnregisterFactory("Microsoft.Data.SqlClient");
     }
 }
