+++ src/Orc.DataAccess/Database/DbDataSourceProviders/MsSqlDbDataSourceProvider.cs
@@ -1,12 +1,6 @@
 ﻿namespace Orc.DataAccess.Database;
 
-using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Common;
-using Microsoft.Win32;
-using Orc.DataAccess.Registry;
-using RegistryKey = Registry.RegistryKey;
+using Services;
 
 [ConnectToProvider("Microsoft.Data.SqlClient")]
 public class MsSqlDbDataSourceProvider : MsSqlDbDataSourceProviderBase
