+++ src/Orc.DataAccess/Registry/Interfaces/IRegistryKey.cs
@@ -1,4 +1,4 @@
-﻿namespace Orc.DataAccess.Registry;
+﻿namespace Orc.DataAccess;
 
 using System;
