+++ src/Orc.DataAccess/Registry/RegistryKey.cs
@@ -1,4 +1,4 @@
-﻿namespace Orc.DataAccess.Registry;
+﻿namespace Orc.DataAccess;
 
 using System;
