+++ src/Orc.DataAccess/Registry/Services/RegistryKeyService.cs
@@ -1,4 +1,4 @@
-﻿namespace Orc.DataAccess.Registry;
+﻿namespace Orc.DataAccess.Services;
 
 using Microsoft.Win32;
 
@@ -6,6 +6,6 @@ internal class RegistryKeyService : IRegistryKeyService
 {
     public IRegistryKey OpenBaseKey(RegistryHive hKey, RegistryView view)
     {
-        return new RegistryKey(Microsoft.Win32.RegistryKey.OpenBaseKey(hKey, view));
+        return new DataAccess.RegistryKey(Microsoft.Win32.RegistryKey.OpenBaseKey(hKey, view));
     }
 }
